@@ -1,40 +1,36 @@
 {
-  "name": "nextjs",
-  "version": "0.1.0",
-  "private": true,
-  "proxy": "https://fraldasdipano.vtexcommercestable.com.br",
+  "name": "mui5",
+  "version": "1.0.0",
+  "description": "This is a [Next.js](https://nextjs.org/) project bootstrapped with [`create-next-app`](https://github.com/vercel/next.js/tree/canary/packages/create-next-app).",
+  "main": "next.config.js",
   "scripts": {
-    "dev": "next dev",
+    "dev": "next dev -p 8000",
     "build": "next build",
-    "start": "next start"
+    "start": "next start -p 8000",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
   "dependencies": {
-    "@date-io/date-fns": "^2.13.1",
-    "@date-io/moment": "^2.13.1",
-    "@emotion/react": "^11.7.1",
-    "@emotion/styled": "^11.6.0",
-    "@material-ui/core": "^5.0.0-beta.5",
-    "@material-ui/data-grid": "^4.0.0-alpha.35",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.60",
-    "@material-ui/styles": "^4.11.4",
-    "@material-ui/types": "^5.1.0",
-    "@mui/icons-material": "^5.2.5",
-    "@mui/lab": "^5.0.0-alpha.69",
-    "@mui/material": "^5.2.6",
-    "@mui/styles": "^5.2.3",
-    "fsevents": "^2.3.2",
-    "install": "^0.13.0",
-    "moment": "^2.29.1",
-    "moment-timezone": "^0.5.34",
-    "next": "^12.0.7",
-    "node-zendesk": "^2.1.0",
-    "npm": "^8.5.1",
-    "pagarme": "^4.20.3",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-draggable": "^4.4.4",
-    "react-scripts": "^5.0.0",
-    "recharts": "^2.1.8"
+    "@date-io/moment": "^3.2.0",
+    "@emotion/react": "^11.14.0",
+    "@emotion/styled": "^11.14.0",
+    "@mui/icons-material": "^5.16.13",
+    "@mui/lab": "^5.0.0-alpha.125",
+    "@mui/material": "^5.16.13",
+    "@mui/x-date-pickers": "^5.0.20",
+    "moment": "^2.30.1",
+    "moment-timezone": "^0.5.46",
+    "next": "^14.2.22",
+    "pagarme": "^4.35.2",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-draggable": "^4.4.6",
+    "recharts": "^2.15.0"
+  },
+  "devDependencies": {
+    "@types/node": "22.10.2",
+    "typescript": "5.7.2"
   }
 }
