@@ -16,28 +16,37 @@ import Paper from '@mui/material/Paper';
 import MenuIcon from '@mui/icons-material/Menu';
 import ChevronLeftIcon from '@mui/icons-material/ChevronLeft';
 import Copyright from "../includes/layoutaccessories/Copyright";
-import Fade from "@material-ui/core/Fade";
-import { Tooltip } from "@material-ui/core";
+import Fade from '@mui/material/Fade';
+import Tooltip from '@mui/material/Tooltip';
 import { PrimaryMenuOptions, SecondaryMenuOptions } from "./MenuOptions";
-import { Notifications, FavoriteOutlined, ErrorOutlined, ArrowDropDown } from "@material-ui/icons";
-import Button from '@mui/material/Button'
+import NotificationsIcon from '@mui/icons-material/Notifications';
+import FavoriteOutlinedIcon from '@mui/icons-material/FavoriteOutlined';
+import ErrorOutlinedIcon from '@mui/icons-material/ErrorOutlined';
+import ArrowDropDownIcon from '@mui/icons-material/ArrowDropDown';
+import Button from '@mui/material/Button';
 import FeedTable from "./FeedTable";
-import FeedBar from '../includes/feedbar/FeedBar'
-import PageTitle from '../includes/layoutaccessories/PageTitle'
+import FeedBar from '../includes/feedbar/FeedBar';
+import PageTitle from '../includes/layoutaccessories/PageTitle';
 import TextField from '@mui/material/TextField';
-import DateAdapter from '@mui/lab/AdapterMoment';
-import LocalizationProvider from '@mui/lab/LocalizationProvider';
-import DesktopDateTimePicker from '@mui/lab/DesktopDateTimePicker';
+import { LocalizationProvider, DesktopDateTimePicker } from '@mui/x-date-pickers';
+import { AdapterMoment } from '@mui/x-date-pickers/AdapterMoment';
 
-const LightTooltip = styled(Tooltip, {})({
-  color: "Ivory",
-  backgroundColor: "transparente",
-  boxShadow: 2,
-  fontSize: 13,
-});
+// ✅ Substituído estilo Tooltip corretamente para MUI 5
+const LightTooltip = styled(({ className, ...props }) => (
+  <Tooltip {...props} classes={{ popper: className }} />
+))(({ theme }) => ({
+  '& .MuiTooltip-tooltip': {
+    color: "Ivory",
+    backgroundColor: "transparent",
+    boxShadow: theme.shadows[2],
+    fontSize: 13,
+  },
+}));
 
+// ✅ Constante para largura do Drawer
 const drawerWidth = 220;
 
+// ✅ Estilização do AppBar
 const AppBar = styled(MuiAppBar, {
   shouldForwardProp: (prop) => prop !== 'open',
 })(({ theme, open }) => ({
@@ -56,6 +65,7 @@ const AppBar = styled(MuiAppBar, {
   }),
 }));
 
+// ✅ Estilização do Drawer
 const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' })(
   ({ theme, open }) => ({
     '& .MuiDrawer-paper': {
@@ -82,8 +92,8 @@ const Drawer = styled(MuiDrawer, { shouldForwardProp: (prop) => prop !== 'open' 
   }),
 );
 
+// ✅ Tema padrão
 const mdTheme = createTheme();
-
 
 export default function DashboardContent(props) {
   const [open, setOpen] = useState(true);
@@ -94,45 +104,39 @@ export default function DashboardContent(props) {
     setOpen(!open);
   };
 
-  // Update Document Title
+  // ✅ Atualiza título da página
   useEffect(() => {
-    props.notificationWhiteList ? document.title = props.notificationWhiteList + ' VIP(s) ' : ''
-    props.notificationAlerts ? document.title = props.notificationAlerts + ' Alerta(s)  ' : ''
-    props.notificationBlackList ? document.title = props.notificationBlackList + ' Blacklist(s)  ' : ''
-  })
+    if (props.notificationWhiteList) document.title = `${props.notificationWhiteList} VIP(s)`;
+    if (props.notificationAlerts) document.title = `${props.notificationAlerts} Alerta(s)`;
+    if (props.notificationBlackList) document.title = `${props.notificationBlackList} Blacklist(s)`;
+  }, [props.notificationWhiteList, props.notificationAlerts, props.notificationBlackList]);
 
   const FeedBarDate = () => {
     return (
       <Box sx={{ ml: 2, mr: 5, display: 'inline' }}>
-        <div>
-          <LocalizationProvider dateAdapter={DateAdapter}>
-            <DesktopDateTimePicker
-              renderInput={(props) => <TextField {...props} size='small' />}
-              label="Início"
-              value={startDate}
-              onChange={(newDate) => {
-                setStartDate(newDate);
-                window.location.reload();
-              }}
-              ampm={false}
-              ampmInClock={false}
-            />
-          </LocalizationProvider>
-        </div>
+        <LocalizationProvider dateAdapter={AdapterMoment}>
+          <DesktopDateTimePicker
+            renderInput={(props) => <TextField {...props} size='small' />}
+            label="Início"
+            value={startDate}
+            onChange={(newDate) => {
+              setStartDate(newDate);
+              window.location.reload();
+            }}
+            ampm={false}
+            ampmInClock={false}
+          />
+        </LocalizationProvider>
       </Box>
-    )
-  }
+    );
+  };
 
   return (
     <ThemeProvider theme={mdTheme}>
       <Box sx={{ display: 'flex' }}>
         <CssBaseline />
         <AppBar position="absolute" open={open} style={{ background: "#d5d5d5" }}>
-          <Toolbar
-            sx={{
-              pr: '24px', // keep right padding when drawer closed
-            }}
-          >
+          <Toolbar sx={{ pr: '24px' }}>
             <IconButton
               edge="start"
               color="inherit"
@@ -145,82 +149,43 @@ export default function DashboardContent(props) {
             >
               <MenuIcon />
             </IconButton>
-
             <PageTitle title='Feed' />
 
-            {/* <FeedBarDate /> */}
-
-
-            {/* Notification Whitelist */}
-            <LightTooltip
-              title="VIP"
-              placement="bottom"
-              arrow
-              interactive
-              TransitionComponent={Fade}
-              TransitionProps={{ timeout: 600 }}
-              aria-label="Cartão"
-            >
-              <IconButton aria-label="whitelist" color="inherit">
+            {/* Notificações */}
+            <LightTooltip title="VIP" arrow>
+              <IconButton color="inherit">
                 <Badge badgeContent={props.notificationWhiteList} color="primary">
-                  <FavoriteOutlined />
+                  <FavoriteOutlinedIcon />
                 </Badge>
               </IconButton>
             </LightTooltip>
-            {/* Notification Blacklist */}
-            <LightTooltip
-              title="Retrições"
-              placement="bottom"
-              arrow
-              interactive
-              TransitionComponent={Fade}
-              TransitionProps={{ timeout: 600 }}
-              aria-label="Cartão"
-            >
-              <IconButton aria-label="blacklist" color="inherit">
-                <Badge
-                  badgeContent={props.notificationBlackList}
-                  color="secondary"
-                >
-                  <ErrorOutlined />
+
+            <LightTooltip title="Restrições" arrow>
+              <IconButton color="inherit">
+                <Badge badgeContent={props.notificationBlackList} color="secondary">
+                  <ErrorOutlinedIcon />
                 </Badge>
               </IconButton>
             </LightTooltip>
-            {/* Notification Alerts */}
-            <LightTooltip
-              title="Alto Risco"
-              placement="bottom"
-              arrow
-              interactive
-              TransitionComponent={Fade}
-              TransitionProps={{ timeout: 600 }}
-              aria-label="Notification"
-            >
-              <IconButton aria-label="notification" color="inherit">
+
+            <LightTooltip title="Alto Risco" arrow>
+              <IconButton color="inherit">
                 <Badge badgeContent={props.notificationAlerts} color="error">
-                  <Notifications />
+                  <NotificationsIcon />
                 </Badge>
               </IconButton>
             </LightTooltip>
 
             <Box sx={{ ml: 1 }}>
               <Button>
-                <ArrowDropDown />
+                <ArrowDropDownIcon />
               </Button>
             </Box>
-
           </Toolbar>
         </AppBar>
 
         <Drawer variant="permanent" open={open}>
-          <Toolbar
-            sx={{
-              display: 'flex',
-              alignItems: 'center',
-              justifyContent: 'flex-end',
-              px: [1],
-            }}
-          >
+          <Toolbar sx={{ display: 'flex', alignItems: 'center', justifyContent: 'flex-end' }}>
             <IconButton onClick={toggleDrawer}>
               <ChevronLeftIcon />
             </IconButton>
@@ -229,44 +194,23 @@ export default function DashboardContent(props) {
           <List>{PrimaryMenuOptions}</List>
           <Divider />
           <List>{SecondaryMenuOptions}</List>
-          <Box sx={{ mt: 20, ml: 10, color: 'Silver', fontSize: 15 }}>
-            AlerteMe
-          </Box>
-          <Box sx={{ ml: 11.5, color: 'Steelblue', fontSize: 10 }}>
-            v. Mui5
-          </Box>
         </Drawer>
 
         {/* FEED TABLE */}
-        <Box
-          component="main"
-          sx={{
-            backgroundColor: (theme) =>
-              theme.palette.mode === 'light'
-                ? theme.palette.grey[100]
-                : theme.palette.grey[900],
-            flexGrow: 1,
-            height: '100vh',
-            overflow: 'auto',
-          }}
-        >
+        <Box component="main" sx={{ flexGrow: 1, height: '100vh', overflow: 'auto' }}>
           <Toolbar />
-
           <FeedBar orders={props.orders} />
-
           <Container maxWidth="lg" sx={{ mt: 1, mb: 4 }}>
             <Grid container spacing={3}>
               <Grid item xs={12}>
-                <Paper sx={{ p: 2, display: 'flex', flexDirection: 'column' }}>
+                <Paper sx={{ p: 2 }}>
                   <FeedTable orders={props.orders} />
                 </Paper>
               </Grid>
             </Grid>
-            {/* <Copyright sx={{ mt: 0.5, pt: 2 }} /> */}
           </Container>
-
         </Box>
       </Box>
-    </ThemeProvider >
+    </ThemeProvider>
   );
 }
