@@ -1,62 +1,57 @@
-import { makeStyles } from "@material-ui/styles";
-import { Timer, Schedule, ScreenShare, ShoppingBasket, MonetizationOn, Cancel, ExposurePlus1 } from "@material-ui/icons";
+import { styled } from "@mui/material/styles";
+import TimerIcon from "@mui/icons-material/Timer";
+import ScheduleIcon from "@mui/icons-material/Schedule";
+import ScreenShareIcon from "@mui/icons-material/ScreenShare";
+import ShoppingBasketIcon from "@mui/icons-material/ShoppingBasket";
+import MonetizationOnIcon from "@mui/icons-material/MonetizationOn";
+import CancelIcon from "@mui/icons-material/Cancel";
+import HelpOutlineIcon from "@mui/icons-material/HelpOutline"; // Ícone padrão para status desconhecido
 
-import ExposurePlus1Icon from "@material-ui/icons/ExposurePlus1";
+// Estilização individual para cada status
+const AguardandoIcon = styled(TimerIcon)(({ theme }) => ({
+  color: "LightGrey",
+}));
 
+const CarenciaIcon = styled(ScheduleIcon)(({ theme }) => ({
+  color: "LightGrey",
+}));
+
+const ManuseioIcon = styled(ScreenShareIcon)(({ theme }) => ({
+  color: "LightGrey",
+}));
+
+const PreparandoIcon = styled(ShoppingBasketIcon)(({ theme }) => ({
+  color: "Tan",
+}));
+
+const FaturadoIcon = styled(MonetizationOnIcon)(({ theme }) => ({
+  color: "LightGrey",
+}));
+
+const CanceladoIcon = styled(CancelIcon)(({ theme }) => ({
+  color: "LightGrey",
+}));
+
+const DefaultIcon = styled(HelpOutlineIcon)(({ theme }) => ({
+  color: "DarkGray",
+}));
+
+// Componente principal com ícones estilizados
 export function IconizeStatus(props) {
-  const classes = useStatusStyles();
-
   switch (props.status) {
     case "payment-pending":
-      return <Timer className={classes.aguardando} fontSize={props.size} />;
+      return <AguardandoIcon fontSize={props.size} />;
     case "window-to-cancel":
-      return (
-        <Schedule className={classes.carencia} fontSize={props.size} />
-      );
+      return <CarenciaIcon fontSize={props.size} />;
     case "ready-for-handling":
-      return (
-        <ScreenShare className={classes.manuseio} fontSize={props.size} />
-      );
+      return <ManuseioIcon fontSize={props.size} />;
     case "handling":
-      return (
-        <ShoppingBasket
-          className={classes.preparando}
-          fontSize={props.size}
-        />
-      );
+      return <PreparandoIcon fontSize={props.size} />;
     case "invoiced":
-      return (
-        <MonetizationOn
-          className={classes.faturado}
-          fontSize={props.size}
-        />
-      );
+      return <FaturadoIcon fontSize={props.size} />;
     case "canceled":
-      return <Cancel className={classes.cancelado} fontSize={props.size} />;
+      return <CanceladoIcon fontSize={props.size} />;
     default:
-      return (
-        <ExposurePlus1 className={classes.qty6} fontSize={props.size} />
-      );
+      return <DefaultIcon fontSize={props.size} />;
   }
 }
-
-const useStatusStyles = makeStyles({
-  aguardando: {
-    color: "LightGrey",
-  },
-  carencia: {
-    color: "LightGrey",
-  },
-  manuseio: {
-    color: "LightGrey",
-  },
-  preparando: {
-    color: "Tan",
-  },
-  faturado: {
-    color: "LightGrey",
-  },
-  cancelado: {
-    color: "LightGrey",
-  },
-});
