@@ -1,12 +1,15 @@
-import TableCell from "@material-ui/core/TableCell";
+import TableCell from "@mui/material/TableCell";
+import PropTypes from 'prop-types';
 
-export const OrderClientCell = (props) => {
+export const OrderClientCell = ({ client }) => {
+  return (
+    <TableCell align="left">
+      {client}
+    </TableCell>
+  );
+};
 
-    const { client } = props 
-
-    return (
-        <TableCell align="left">
-          {client}
-        </TableCell>
-    )
-}
+// Validação de Propriedades (Boa Prática)
+OrderClientCell.propTypes = {
+  client: PropTypes.string.isRequired,
+};
