@@ -1,12 +1,15 @@
-import TableCell from "@material-ui/core/TableCell";
+import TableCell from "@mui/material/TableCell";
+import PropTypes from 'prop-types';
 
-export const OrderCreateDateCell = (props) => {
+export const OrderCreateDateCell = ({ createDate }) => {
+  return (
+    <TableCell align="center" size="small">
+      {createDate}
+    </TableCell>
+  );
+};
 
-    const { createDate } = props 
-
-    return (
-        <TableCell align="center" size="small">
-          {createDate}
-        </TableCell>
-    )
-}
+// Validação de Propriedades (Boa Prática)
+OrderCreateDateCell.propTypes = {
+  createDate: PropTypes.string.isRequired,
+};
