@@ -1,28 +1,29 @@
 import { useState } from "react";
-import { TableCell, IconButton, Tooltip, Fade } from "@material-ui/core";
-import { withStyles } from '@mui/styles';
+import { TableCell, IconButton, Tooltip, Fade } from "@mui/material";
+import { styled } from "@mui/material/styles"; // Correta substituição para `withStyles`
 import { StyledBadgeHist } from "../badgezation/StyledBadge";
 import { IconizePurchaseProfile } from "../iconization/IconizePurchaseProfile";
 import setCurrency from "../../../../backend/utils/setCurrency";
-import HistoryWindow from '../dialogs/HistoryWindow'
+import HistoryWindow from "../dialogs/HistoryWindow";
 
-const LightTooltip = withStyles((theme) => ({
-  tooltip: {
+// Estilização do Tooltip usando `styled`
+const LightTooltip = styled(({ className, ...props }) => (
+  <Tooltip {...props} classes={{ popper: className }} />
+))(({ theme }) => ({
+  '& .MuiTooltip-tooltip': {
     backgroundColor: theme.palette.action.active,
     color: "Ivory",
     boxShadow: theme.shadows[2],
     fontSize: 13,
   },
-}))(Tooltip);
+}));
 
-<StyledBadgeHist />;
-
-
+// Componente Principal
 export const OrderHistoryCell = (props) => {
   const { orderDetail } = props;
   const [isOpen, setIsOpen] = useState(false);
 
-  const handleCellClick = (e) => setIsOpen((prevState) => !prevState);
+  const handleCellClick = () => setIsOpen((prevState) => !prevState);
 
   return (
     <>
@@ -30,8 +31,7 @@ export const OrderHistoryCell = (props) => {
         <LightTooltip
           title={
             orderDetail.valuePurchase
-              ? "Histórico total de compras de  " +
-              setCurrency(orderDetail.valuePurchase)
+              ? "Histórico total de compras de " + setCurrency(orderDetail.valuePurchase)
               : "Primeira compra"
           }
           placement="top-end"
@@ -42,13 +42,11 @@ export const OrderHistoryCell = (props) => {
           aria-label="Order History"
         >
           <IconButton>
-            {orderDetail.qtyPurchase > 0 ? (
+            {orderDetail.qtyPurchase > 0 && (
               <StyledBadgeHist
                 badgeContent={orderDetail.qtyPurchase}
                 max={999}
-              ></StyledBadgeHist>
-            ) : (
-              ""
+              />
             )}
             <IconizePurchaseProfile
               qtyPurchase={orderDetail.qtyPurchase}
