@@ -1,40 +1,37 @@
-import { React, useState, useEffect } from "react";
-import TableCell from "@material-ui/core/TableCell";
-import IconButton from "@material-ui/core/IconButton";
-import Tooltip from "@material-ui/core/Tooltip";
-import { Fade } from "@material-ui/core";
-import { withStyles } from '@mui/styles';
+import React, { useState, useEffect } from "react";
+import TableCell from "@mui/material/TableCell";
+import IconButton from "@mui/material/IconButton";
+import Tooltip from "@mui/material/Tooltip";
+import Fade from "@mui/material/Fade";
+import { styled } from "@mui/material/styles";
+
 import { IconizeRiskLevel } from "../iconization/IconizeRiskLevel";
 import { StyledBadge } from "../badgezation/StyledBadge";
 import { StyledBadgeRisk } from "../badgezation/StyledBadge";
 import { StyledBadgeWarning } from "../badgezation/StyledBadge";
-import ScoreWindow from "../dialogs/ScoreWindow"
+import ScoreWindow from "../dialogs/ScoreWindow";
 
-
-const LightTooltip = withStyles((theme) => ({
-  tooltip: {
+// Estilização do Tooltip com styled
+const LightTooltip = styled(({ className, ...props }) => (
+  <Tooltip {...props} classes={{ popper: className }} />
+))(({ theme }) => ({
+  "& .MuiTooltip-tooltip": {
     backgroundColor: theme.palette.action.active,
     color: "Ivory",
     boxShadow: theme.shadows[2],
     fontSize: 13,
   },
-}))(Tooltip);
-
-<>
-  <StyledBadge />
-  <StyledBadgeRisk />
-  <StyledBadgeWarning />
-</>;
+}));
 
 export const OrderScoreCell = (props) => {
   const { orderDetail } = props;
   const [isOpen, setIsOpen] = useState(false);
 
-  const handleCellClick = (e) => setIsOpen((prevState) => !prevState);
+  const handleCellClick = () => setIsOpen((prevState) => !prevState);
 
   return (
     <>
-      <TableCell onClick={handleCellClick} lign="left">
+      <TableCell onClick={handleCellClick} align="left">
         <LightTooltip
           title={`Risco ${orderDetail.scoreDesc}`}
           placement="top-end"
@@ -49,17 +46,17 @@ export const OrderScoreCell = (props) => {
               <StyledBadgeRisk
                 badgeContent={orderDetail.score}
                 max={999}
-              ></StyledBadgeRisk>
+              />
             ) : orderDetail.score > 60 ? (
               <StyledBadgeWarning
                 badgeContent={orderDetail.score}
                 max={999}
-              ></StyledBadgeWarning>
+              />
             ) : (
               <StyledBadge
                 badgeContent={orderDetail.score}
                 max={999}
-              ></StyledBadge>
+              />
             )}
             <IconizeRiskLevel
               riskLevel={orderDetail.scoreDesc}
@@ -68,7 +65,12 @@ export const OrderScoreCell = (props) => {
           </IconButton>
         </LightTooltip>
       </TableCell>
-      {isOpen && <ScoreWindow windowState={handleCellClick} orderDetail={orderDetail} />}
+      {isOpen && (
+        <ScoreWindow
+          windowState={handleCellClick}
+          orderDetail={orderDetail}
+        />
+      )}
     </>
   );
 };
