@@ -1,15 +1,16 @@
-import TableHead from "@material-ui/core/TableHead";
-import TableRow from "@material-ui/core/TableRow";
-import TableCell from "@material-ui/core/TableCell";
+import TableHead from "@mui/material/TableHead";
+import TableRow from "@mui/material/TableRow";
+import TableCell from "@mui/material/TableCell";
 import { styled } from '@mui/material/styles';
 
-
-const StyledTableCell = styled(TableCell, {})({
+// Estilização do TableCell utilizando styled
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
   color: "Gray",
-  backgroundColor: "Snow",
-  boxShadow: 2,
+  backgroundColor: theme.palette.grey[100], // Usar o sistema de cores do Material-UI
+  boxShadow: theme.shadows[1], // Aplicar sombra conforme o padrão Material-UI
   fontSize: 11,
-});
+  fontWeight: 'bold',
+}));
 
 export const OrderTableHeadHist = () => {
   return (
