@@ -1,6 +1,8 @@
-import { createTheme } from '@material-ui/core/styles';
-import { ptBR } from '@mui/material/locale'
-import { red } from "@material-ui/core/colors";
+import { createTheme } from '@mui/material/styles';
+import { red } from '@mui/material/colors';
+import { ptBR } from '@mui/material/locale';
+
+
 
 // Create a theme instance.
 const theme = createTheme({
