@@ -1,5 +1,279 @@
 import titleCase from "../src/backend/utils/titleCase";
-const BLACK_LIST = [{
+const BLACK_LIST = [
+    {
+        blackedName: "Tamara Talifa Alves Rodrigues",
+        blackedEmail: "tiagofernandeshungen@gmail.com",
+        blackedCPF: "04917906199",
+        blackedPhone: "+5565992543162",
+        blackedPostalCode: "78052-320",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Suspeita de Fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Gabriel Silva",
+        blackedEmail: "gabrielcardozo.2030@icloud.com",
+        blackedCPF: "13557135799",
+        blackedPhone: "+5512988530365",
+        blackedPostalCode: "26176-285",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Suspeita de Fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Debora Santos",
+        blackedEmail: "deborasilva121094@gmail.com",
+        blackedCPF: "16586045711",
+        blackedPhone: "+5521969556504",
+        blackedPostalCode: "26187-210",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Suspeita de Fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Oseias Belmiro",
+        blackedEmail: "oseiasbelmiro4@gmail.com",
+        blackedCPF: "18699652720",
+        blackedPhone: "+5521966864325",
+        blackedPostalCode: "26187-210",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Suspeita de Fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Francisco Santos",
+        blackedEmail: "elivanfrancisco30@gmail.com",
+        blackedCPF: "34315004200",
+        blackedPhone: "+5592985242444",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude confirmada",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Vivian Ataides Braz",
+        blackedEmail: "vivianataidesbraz@gmail.com",
+        blackedCPF: "89147367172",
+        blackedPhone: "+5594984106002",
+        blackedPostalCode: "68513-400",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude confirmada",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "",
+        blackedEmail: "luiz.henrique@wedigi.com.br",
+        blackedCPF: "",
+        blackedPhone: "",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Renata Amaral",
+        blackedEmail: "renatinhaamaral89@gmail.com",
+        blackedCPF: "87751976420",
+        blackedPhone: "",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "Visa 498442****9222",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude Confirmada",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Rui Motomu",
+        blackedEmail: "rmiittoo@gmail.com",
+        blackedCPF: "17988846825",
+        blackedPhone: "+5584981869053",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude confirmada",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Katia Souza",
+        blackedEmail: "katiamuta@gmail.com",
+        blackedCPF: "21544177879",
+        blackedPhone: "+5519981009927",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Apenas monitorando pois não respondeu nossos contatos",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Naiara Martins Dos Reis",
+        blackedEmail: "naiarareis484@gmail.com",
+        blackedCPF: "01690893206",
+        blackedPhone: "+5594991924312",
+        blackedPostalCode: "68509-700",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Suspeita de Fraude. Compras sucessivas",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Ludmila Rocha",
+        blackedEmail: "ludyaugusta27@gmail.com",
+        blackedCPF: "02087810133",
+        blackedPhone: "+5562982431664",
+        blackedPostalCode: "74215-110",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude em potencial. Portador do cartão não reconhece a compra.",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "leticia an",
+        blackedEmail: "onlineon659@gmail.com",
+        blackedCPF: "38013124886",
+        blackedPhone: "+5513997320803",
+        blackedPostalCode: "22775-003",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "",
+        blackedEmail: "",
+        blackedCPF: "",
+        blackedPhone: "",
+        blackedPostalCode: "22775-003",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude por BOT",
+        isBlackListed: false,
+    },
+    {
+        blackedName: "Caroline Alfama Fernandes Simões",
+        blackedEmail: "carol.alfama@gmail.com",
+        blackedCPF: "05910603701",
+        blackedPhone: "+5521968557132",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Solicitação da Cris para blacklist",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Debora Cassia",
+        blackedEmail: "deboradecassia135@gmail.com",
+        blackedCPF: "17448462865",
+        blackedPhone: "+5511977373017",
+        blackedPostalCode: "13045-310",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "Mastercard 515590****6508",
+        blackedCardCountry: "",
+        blackedDescription: "Várias transações negadas pela Pagarme",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "ANA SANTOS",
+        blackedEmail: "anaameliadossantos12@outlook.com",
+        blackedCPF: "47939349864",
+        blackedPhone: "+5512991876522",
+        blackedPostalCode: "11629-248",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "Mastercard 523431****2220",
+        blackedCardCountry: "",
+        blackedDescription: "Chargeback por fraude",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Cleidiane Matos",
+        blackedEmail: "cleidianebrito@yahoo.com.br",
+        blackedCPF: "81754906249",
+        blackedPhone: "+559199023097",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "550209****8806",
+        blackedCardCountry: "",
+        blackedDescription: "Muitas compras seguidas sem autenticação",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Nicolle Laffi",
+        blackedEmail: "nicolle.laffi@gmail.com",
+        blackedCPF: "6850465864",
+        blackedPhone: "+5511999378878",
+        blackedPostalCode: "",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "",
+        blackedCardCountry: "",
+        blackedDescription: "Chargebck - Cliente não reconheceu a compra",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Severino Soares",
+        blackedEmail: "severinosoares1601@gmail.com",
+        blackedCPF: "60712201491",
+        blackedPhone: "+5565981023631",
+        blackedPostalCode: "78135-120",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "Elo 650495****3100",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude - Richard",
+        isBlackListed: true,
+    },
+    {
+        blackedName: "Daniel Pereira De Souza Souza",
+        blackedEmail: "richardandredasilva19@gmail.com",
+        blackedCPF: "05861290733",
+        blackedPhone: "+5565981215978",
+        blackedPostalCode: "78135-120",
+        blackedState: "",
+        blackedCity: "",
+        blackedCreditCard: "Elo 650495****3100",
+        blackedCardCountry: "",
+        blackedDescription: "Fraude - Richard",
+        isBlackListed: true,
+    },
+    {
         blackedName: "Julio Carnicelli",
         blackedEmail: "",
         blackedCPF: "22284392822",
@@ -868,9 +1142,9 @@ export function isBlackListed(
     );
     let isBlackedCardCountry = BLACK_LIST.find(
         (bl) =>
-        bl.blackedCardCountry == cardCountry &&
-        cardCountry > " " &&
-        bl.isBlackListed
+            bl.blackedCardCountry == cardCountry &&
+            cardCountry > " " &&
+            bl.isBlackListed
     );
 
     let isBlacked = false;
