@@ -1,63 +1,69 @@
-import React from "react";
-import { useState, useEffect } from 'react'
-import FeedTableDialog from './FeedTableDialog'
-import FeedTableDialogTitle from './FeedTableDialogTitle'
-import { makeStyles, withStyles } from "@mui/styles";
-import { Paper, Grid, Box, DialogContent, DialogContentText } from '@mui/material'
-import { Table, TableBody, TableCell, TableContainer, TableHead, TableRow } from "@mui/material";
-import Draggable from 'react-draggable';
+import React, { useState, useEffect } from "react";
+import FeedTableDialog from "./FeedTableDialog";
+import FeedTableDialogTitle from "./FeedTableDialogTitle";
+import {
+  Paper,
+  Grid,
+  Box,
+  DialogContent,
+  DialogContentText,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+} from "@mui/material";
+import { styled } from "@mui/material/styles";
+
 import formatTZOrderDate from "../../../../backend/utils/formatTZOrderDate";
 import setCurrency from "../../../../backend/utils/setCurrency";
 
-const useStyles = makeStyles((theme) => ({
-  title: {
-    flex: 1,
+// ✅ Estilização com `styled`
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  "&.MuiTableCell-head": {
+    backgroundColor: "#d5d5d5",
+    fontSize: 14,
+    fontWeight: "bold",
   },
-  table: {
-    minWidth: 650,
+  "&.MuiTableCell-body": {
+    fontSize: 14,
+    color: theme.palette.text.primary,
   },
-  color: "Grey",
 }));
 
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: "#d5d5d5",
+const StyledTableRow = styled(TableRow)(({ theme }) => ({
+  "&:nth-of-type(odd)": {
+    backgroundColor: theme.palette.action.hover,
   },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    "&:nth-of-type(odd)": {
-    },
-  },
-}))(TableRow);
-
+}));
 
 const IncompleteOrdersWindow = (props) => {
-  const classes = useStyles();
   const { orderDetail, windowState } = props;
-  const [cpf, setCpf] = useState(orderDetail.history[0].cpf);
-  const [clientName, setClientName] = useState(orderDetail.cliente);
+  const [cpf, setCpf] = useState(orderDetail.history[0]?.cpf || "");
+  const [clientName, setClientName] = useState(orderDetail.cliente || "");
   const [open, setOpen] = useState(true);
-  const [historyItems, setHistoryItems] = useState([{}]);
+  const [historyItems, setHistoryItems] = useState([]);
   const [loading, setLoading] = useState(true);
   const [item, setItem] = useState({});
 
-  useEffect(async () => {
-    // setClientName(orderDetail.cliente);
-    // const url = `/api/incomplete/${clientName}`;
-    setCpf(orderDetail.history[0].cpf);
-    const url = `/api/incompletebycpf/${cpf}`;
-    const response = await fetch(url, { method: "GET" });
-    const data = await response.json();
-    console.log("Data History=> ", data.history);
-    setHistoryItems(data.history);
-    getCurrentOrder();
-    setLoading(false);
-  }, []);
+  useEffect(() => {
+    const fetchHistory = async () => {
+      try {
+        const url = `/api/incompletebycpf/${cpf}`;
+        const response = await fetch(url, { method: "GET" });
+        const data = await response.json();
+        setHistoryItems(data.history || []);
+        getCurrentOrder();
+        setLoading(false);
+      } catch (error) {
+        console.error("Erro ao buscar dados incompletos:", error);
+        setLoading(false);
+      }
+    };
+
+    if (cpf) fetchHistory();
+  }, [cpf]);
 
   const getCurrentOrder = () => {
     const item = {
@@ -68,7 +74,7 @@ const IncompleteOrdersWindow = (props) => {
       list: orderDetail.giftId,
       payment: orderDetail.payMethod,
       creditCard: orderDetail.creditCard,
-      installments: orderDetail.history[0].parcelas,
+      installments: orderDetail.history[0]?.parcelas || "",
       tid: "",
       statusDescription: orderDetail.statusDescription,
       reason: "-",
@@ -79,15 +85,12 @@ const IncompleteOrdersWindow = (props) => {
   return (
     <FeedTableDialog windowState={windowState}>
       <FeedTableDialogTitle orderDetail={orderDetail} />
-      {/* CONTENT */}
-      <DialogContent sx={{ bgcolor: "WhiteSmoke" }} >
+
+      {/* ✅ Conteúdo */}
+      <DialogContent sx={{ bgcolor: "WhiteSmoke" }}>
         <DialogContentText>
           <TableContainer component={Paper}>
-            <Table
-              className={classes.table}
-              size="small"
-              aria-label="a dense table"
-            >
+            <Table size="small" aria-label="Pedidos Incompletos">
               <TableHead>
                 <TableRow>
                   <StyledTableCell>Data</StyledTableCell>
@@ -102,12 +105,20 @@ const IncompleteOrdersWindow = (props) => {
               </TableHead>
               <TableBody>
                 {loading ? (
-                  "Buscando dados históricos..."
+                  <TableRow>
+                    <TableCell colSpan={8} align="center">
+                      Buscando dados históricos...
+                    </TableCell>
+                  </TableRow>
                 ) : (
                   <>
                     <FormatDetailLine detailLine={item} index={item.orderId} />
                     {historyItems.map((item, orderId) => (
-                      <FormatDetailLine detailLine={item} index={orderId} />
+                      <FormatDetailLine
+                        detailLine={item}
+                        index={orderId}
+                        key={orderId}
+                      />
                     ))}
                   </>
                 )}
@@ -117,38 +128,36 @@ const IncompleteOrdersWindow = (props) => {
         </DialogContentText>
       </DialogContent>
     </FeedTableDialog>
-  )
-}
+  );
+};
 
 const FormatDetailLine = (props) => {
   const { detailLine, index } = props;
 
   return (
-    <>
-      <StyledTableRow key={index}>
-        <StyledTableCell component="th" scope="row">
-          {detailLine.date}
-        </StyledTableCell>
-        <StyledTableCell align="right">
-          {detailLine.orderId.substr(1, 6)}
-        </StyledTableCell>
-        <StyledTableCell align="center">{detailLine.items}</StyledTableCell>
-        <StyledTableCell align="right">
-          {setCurrency(detailLine.value)}
-        </StyledTableCell>
-        <StyledTableCell align="center">
-          {detailLine.creditCard ? detailLine.creditCard : ""}
-        </StyledTableCell>
-        <StyledTableCell align="center">
-          {detailLine.installments}
-        </StyledTableCell>
-        <StyledTableCell align="left">
-          {detailLine.statusDescription}
-        </StyledTableCell>
-        <StyledTableCell align="center">{detailLine.reason}</StyledTableCell>
-      </StyledTableRow>
-    </>
+    <StyledTableRow key={index}>
+      <StyledTableCell component="th" scope="row">
+        {formatTZOrderDate(detailLine.date).substr(0, 10)}
+      </StyledTableCell>
+      <StyledTableCell align="right">
+        {detailLine.orderId?.substr(1, 6) || ""}
+      </StyledTableCell>
+      <StyledTableCell align="center">{detailLine.items}</StyledTableCell>
+      <StyledTableCell align="right">
+        {setCurrency(detailLine.value)}
+      </StyledTableCell>
+      <StyledTableCell align="center">
+        {detailLine.creditCard || ""}
+      </StyledTableCell>
+      <StyledTableCell align="center">
+        {detailLine.installments || ""}
+      </StyledTableCell>
+      <StyledTableCell align="left">
+        {detailLine.statusDescription}
+      </StyledTableCell>
+      <StyledTableCell align="center">{detailLine.reason}</StyledTableCell>
+    </StyledTableRow>
   );
 };
 
-export default IncompleteOrdersWindow
+export default IncompleteOrdersWindow;
