@@ -1,90 +1,106 @@
-import React, { useState, useEffect } from 'react'
-import { makeStyles, withStyles } from '@mui/styles'
-import { Table, TableHead, TableRow, TableBody, TableCell, TableContainer, Paper, DialogContentText, DialogContent } from '@mui/material'
-import FeedTableDialog from './FeedTableDialog'
-import FeedTableDialogTitle from './FeedTableDialogTitle'
-import setCurrency from "../../../../backend/utils/setCurrency";
+import React, { useState, useEffect } from 'react';
+import {
+  Table,
+  TableHead,
+  TableRow,
+  TableBody,
+  TableCell,
+  TableContainer,
+  Paper,
+  DialogContentText,
+  DialogContent,
+} from '@mui/material';
+import { styled } from '@mui/material/styles';
 
-const useStyles = makeStyles((theme) => ({
-  title: {
-    // marginLeft: theme.spacing(2),
-    flex: 1,
+import FeedTableDialog from './FeedTableDialog';
+import FeedTableDialogTitle from './FeedTableDialogTitle';
+import setCurrency from '../../../../backend/utils/setCurrency';
+
+// ✅ Estilização com `styled`
+const StyledTableCell = styled(TableCell)(({ theme }) => ({
+  '&.MuiTableCell-head': {
+    backgroundColor: '#d5d5d5',
+    fontSize: 14,
+    fontWeight: 'bold',
   },
-  color: "Grey",
+  '&.MuiTableCell-body': {
+    fontSize: 14,
+    color: theme.palette.text.primary,
+  },
 }));
 
-const StyledTableCell = withStyles((theme) => ({
-  head: {
-    backgroundColor: "#d5d5d5",
+const StyledTableRow = styled(TableRow)(({ theme }) => ({
+  '&:nth-of-type(odd)': {
+    backgroundColor: theme.palette.action.hover,
   },
-  body: {
-    fontSize: 14,
-  },
-}))(TableCell);
-
-const StyledTableRow = withStyles((theme) => ({
-  root: {
-    "&:nth-of-type(odd)": {
-      // backgroundColor: theme.palette.action.hover,
-    },
-  },
-}))(TableRow);
+}));
 
 const ItemsWindows = (props) => {
-  const classes = useStyles()
-  const { orderDetail, windowState } = props
-  const [orderId, setOrder] = useState(orderDetail.orderId);
-  const [historyItems, setHistoryItems] = useState([{}]);
+  const { orderDetail, windowState } = props;
+  const [orderId, setOrder] = useState(orderDetail?.orderId || '');
+  const [historyItems, setHistoryItems] = useState([]);
   const [loading, setLoading] = useState(true);
 
-  useEffect(async () => {
-    setOrder(orderDetail.orderId);
-    const url = `/api/items/v${orderId}frdp-01`;
-    const response = await fetch(url, { method: "GET" });
-    const data = await response.json();
-    setHistoryItems(data.history);
-    setLoading(false);
-  }, []);
+  useEffect(() => {
+    const fetchHistory = async () => {
+      try {
+        setOrder(orderDetail.orderId);
+        const url = `/api/items/v${orderId}frdp-01`;
+        const response = await fetch(url, { method: 'GET' });
+        const data = await response.json();
+        setHistoryItems(data.history || []);
+        setLoading(false);
+      } catch (error) {
+        console.error('Erro ao buscar dados históricos:', error);
+        setLoading(false);
+      }
+    };
 
+    if (orderId) fetchHistory();
+  }, [orderId, orderDetail]);
 
   return (
     <FeedTableDialog windowState={windowState}>
       <FeedTableDialogTitle orderDetail={orderDetail} />
-      <DialogContent sx={{ bgcolor: "WhiteSmoke" }} >
+      {/* ✅ Conteúdo */}
+      <DialogContent sx={{ bgcolor: 'WhiteSmoke' }}>
         <DialogContentText>
           <TableContainer component={Paper}>
-            <Table
-              className={classes.table}
-              size="small"
-              aria-label="a dense table"
-            >
+            <Table size="small" aria-label="Detalhes dos Itens">
               <TableHead>
                 <TableRow>
-                  <StyledTableCell align="left">Refência</StyledTableCell>
+                  <StyledTableCell align="left">Referência</StyledTableCell>
                   <StyledTableCell align="left">SKU</StyledTableCell>
                   <StyledTableCell align="left">Nome</StyledTableCell>
                   <StyledTableCell align="right">Quantidade</StyledTableCell>
                   <StyledTableCell align="right">Valor</StyledTableCell>
-                  {/* <StyledTableCell align="left">Produto</StyledTableCell> */}
                 </TableRow>
               </TableHead>
               <TableBody>
                 {loading ? (
-                  "Buscando dados históricos..."
+                  <TableRow>
+                    <TableCell colSpan={5} align="center">
+                      Buscando dados históricos...
+                    </TableCell>
+                  </TableRow>
                 ) : (
                   <>
                     {historyItems.map((item, itemId) => (
                       <StyledTableRow key={itemId}>
-                        <StyledTableCell align="left">{item.refId}</StyledTableCell>
                         <StyledTableCell align="left">
-                          {item.productId}
+                          {item.refId || '-'}
                         </StyledTableCell>
-                        <StyledTableCell align="left">{item.name}</StyledTableCell>
-                        <StyledTableCell align="right">
-                          {item.quantity}
+                        <StyledTableCell align="left">
+                          {item.productId || '-'}
+                        </StyledTableCell>
+                        <StyledTableCell align="left">
+                          {item.name || '-'}
                         </StyledTableCell>
                         <StyledTableCell align="right">
-                          {setCurrency(item.price)}
+                          {item.quantity || '-'}
+                        </StyledTableCell>
+                        <StyledTableCell align="right">
+                          {setCurrency(item.price || 0)}
                         </StyledTableCell>
                       </StyledTableRow>
                     ))}
@@ -96,7 +112,7 @@ const ItemsWindows = (props) => {
         </DialogContentText>
       </DialogContent>
     </FeedTableDialog>
-  )
-}
+  );
+};
 
-export default ItemsWindows
+export default ItemsWindows;
