@@ -1,31 +1,31 @@
-import TableCell from "@material-ui/core/TableCell";
-import IconButton from "@material-ui/core/IconButton";
-import Fade from "@material-ui/core/Fade";
-import Tooltip from "@material-ui/core/Tooltip";
-import { withStyles } from '@mui/styles';
-import { StyledBadgeIncompleteOrders } from "../badgezation/StyledBadge";
-import { StyledBadgeCoupon } from "../badgezation/StyledBadge"
-import { IconizePaymentOption } from "../iconization/IconizePaymentOption";
 import { useState } from "react";
+import TableCell from "@mui/material/TableCell";
+import IconButton from "@mui/material/IconButton";
+import Fade from "@mui/material/Fade";
+import Tooltip from "@mui/material/Tooltip";
+import { styled } from "@mui/material/styles";
+
+import { StyledBadgeIncompleteOrders } from "../badgezation/StyledBadge";
+import { IconizePaymentOption } from "../iconization/IconizePaymentOption";
 import IncompleteOrdersWindow from "../dialogs/IncompleteOrdersWindow";
 
-const LightTooltip = withStyles((theme) => ({
-  tooltip: {
+// Substituindo `withStyles` por `styled`
+const LightTooltip = styled(({ className, ...props }) => (
+  <Tooltip {...props} classes={{ popper: className }} />
+))(({ theme }) => ({
+  "& .MuiTooltip-tooltip": {
     backgroundColor: theme.palette.action.active,
     color: "Ivory",
     boxShadow: theme.shadows[2],
     fontSize: 13,
   },
-}))(Tooltip);
-
-<StyledBadgeIncompleteOrders />;
+}));
 
 export const OrderPaymentCell = (props) => {
-  // const { creditCard, incompleteOrders, paymentMethod } = props;
   const { orderDetail } = props;
   const [isOpen, setIsOpen] = useState(false);
 
-  const handleCellClick = (e) => setIsOpen((prevState) => !prevState);
+  const handleCellClick = () => setIsOpen((prevState) => !prevState);
 
   return (
     <>
@@ -43,11 +43,8 @@ export const OrderPaymentCell = (props) => {
             {orderDetail.incompleteOrders > 0 ? (
               <StyledBadgeIncompleteOrders
                 badgeContent={orderDetail.incompleteOrders}
-              ></StyledBadgeIncompleteOrders>
-            ) : (
-              " "
-            )}
-
+              />
+            ) : null}
             <IconizePaymentOption
               payMethod={orderDetail.payMethod}
               size="default"
@@ -55,7 +52,12 @@ export const OrderPaymentCell = (props) => {
           </IconButton>
         </LightTooltip>
       </TableCell>
-      {isOpen && <IncompleteOrdersWindow windowState={handleCellClick} orderDetail={orderDetail} />}
+      {isOpen && (
+        <IncompleteOrdersWindow
+          windowState={handleCellClick}
+          orderDetail={orderDetail}
+        />
+      )}
     </>
   );
 };
